refactor v2 context tests
